interpreter/requirements.txt
matplotlib
pandas
python-magic
